Backgrounded invokeDelayedProcessing
Doing this intensive method causes upwards of 3 full seconds of UI
freeze. There's nothing in there except presenting the UIAlertView
that needs to happen on the main thread. Wrapping most of the
function in a dispatch_async fixes a freeze on startup or wake
that our team had experienced a lot.