Support Overriding CFBundleVersion
Adds the ability for the user to specify the plist value for key
“HockeyAppVersion” which will override the CFBundleVersion HockeyApp
uses when checking for updates. This is for users who, for whatever
reason, are unable to change their version numbering in CFBundleVersion
and don’t increment it for every build.

This DOES NOT add support in the anything other than checking for
updates and has not been verified for the other cases.