Don't intentionally release identificationCompletion block
The previous state prevented that the completion block was properly called everytime a user tries to enter their authentication credentials (e.g. email, etc.) into BITAuthenticationViewController.
That meant that if the user didn't get it right the first time, there was no way to react to the continuous fails or the eventual correct attempt. This fixes #273.