Fix bug: use correct cache folder for attachment
The attachment filename is encoded with the full file path containing the cache root folder.
On the next app start it is restored, but the cache root folder usually changes between app sessions, so the attachment filename will be invalid.
The solution always sets the correct file path (even if it is already correct).